Made a small change to the third para pointing out appropriateness of major updates for changes between organizations/producers.
Made a small change to the third para of the "When to use each?" section again pointing out appropriateness of major updates for changes between organizations/producers.

Made more substantial changes to Revocation section to clarify that revocation does not revoke all previous versions. If a @timestamp is provided, it revokes just a single version. If no @timestamp is provided then it revokes all versions with the referenced @id.
Modified the sample XML to reflect the above and corrected the @id in the reference to be @idref