A few tweaks and additions throughout.
I think there should be a couple of other changes made to the XML sample snippet but I did not make the changes to avoid risk of it getting out of sync with the Python code.
John, could you please take a look at these and make the changes as appropriate.
- The @negate attribute on the Indicator is unnecessary as it is False by default
- The "xsi:type='ttp:TTPType' " and "version="1.1"" on the Indicated_TTP/TTP reference entry are inappropriate. These are defined when specifying a TTP but not when referencing one. All you need here is the @idref