How's this for a "STIX Community" landing page?
@jonathanbaker, @johnwunder, @sbarnum, @bschmoker 

As part of this I recommend adding "Community" as a menu item in the Main Site Navigation to the right of "About".  I also recommend "Supporters" be a drop-down item under the new "Community" menu item, so visitors can go directly to that page from the main site nav without having to scroll down the Community page to find it. 

Also, I wasn't sure if the OASIS Transition section should have a time frame mentioned, or just leave it to the announcement & faq? I used May 2015 in two places above so it was in there but not too specific in case the actual dates move at all.

What do you think?