Create Python API links for multiple versions.
Both 1.1.1 and 1.2 versions of the data model should have links to
Python documentation. 1.1.1 is hard-coded to the last release of
python-stix 1.1.1.X  (i.e. 1.1.1.5), while 1.2 points to the "stable"
release. Though this is currently also 1.1.1.5, once 1.2.0.0 is
released, "stable" will point there.

The actual results of running "rake regenerate" will be in a separate
commit.