Added TAXII to page title so it's STIX/TAXII
As I recall it was decided that there would be 3 pages, a combined STIX/TAXII page that would also be linked to from the TAXII github.io site, and then their own individual pages on the CybOX and MAEC and github.io sites?  I am working on the TAXII github.io site updates now, so as part of that I've added TAXII to the title of this page and the intro text.